refactor: separated controllers per domain.
